Project Simple Chatty Bot
Objective
Your bot can ask anything you want, but there are two rules for your output:

the line with the test should end with the question mark character;
an option starts with a digit followed by the dot (1., 2., 3., 4.)
If a user enters an incorrect answer, the bot may print a message: 
Please, try again.
The program should stop on the correct answer and print Congratulations, have a nice day! at the end.